fixes for latest updates to master
