+++ test/market/actions/load-basic-market-test.js
@@ -1,0 +1,5 @@
+import {assert} from 'chai';
+
+describe(`modules/market/actions/load-basic-market.js`, () => {
+  it(`should load a basic market`);
+});
